refactor: keyboard
-Separate calculator component into two components one for display and another for keys. -Use calculator keys component as a base for keyboard component to be used for lessons and speed test. -Hide calculator keys on small screens.
